Fix travis tests

+++ net/src/mbedtls/test/data_loader.c
@@ -19,6 +19,8 @@
 
 #include <unistd.h>
 
+#include <sys/stat.h>
+
 AVS_UNIT_TEST(backend_mbedtls, chain_loading_from_file) {
     mbedtls_x509_crt *chain = NULL;
 
@@ -49,19 +51,48 @@ AVS_UNIT_TEST(backend_mbedtls, chain_loading_from_path) {
     mbedtls_x509_crt_free(chain);
 
     // Empty directory case.
-    char name[] = "/tmp/empty-XXXXXX";
-    (void) mkdtemp(name);
-    const avs_net_trusted_cert_info_t empty_dir =
-            avs_net_trusted_cert_info_from_path(name);
-    int retval = _avs_net_mbedtls_load_ca_certs(&chain, &empty_dir);
-    (void) rmdir(name);
-    AVS_UNIT_ASSERT_SUCCESS(retval);
+    {
+        char name[] = "/tmp/empty-XXXXXX";
+        AVS_UNIT_ASSERT_NOT_NULL(mkdtemp(name));
+        const avs_net_trusted_cert_info_t empty_dir =
+                avs_net_trusted_cert_info_from_path(name);
+        int retval = _avs_net_mbedtls_load_ca_certs(&chain, &empty_dir);
+        (void) rmdir(name);
+        AVS_UNIT_ASSERT_SUCCESS(retval);
+    }
 
-    // Directory without permissions - hopefully nobody runs tests as root.
-    const avs_net_trusted_cert_info_t no_permissions_dir =
-            avs_net_trusted_cert_info_from_path("/root");
-    AVS_UNIT_ASSERT_FAILED(
-            _avs_net_mbedtls_load_ca_certs(&chain, &no_permissions_dir));
+    // Directory without permissions.
+    {
+        uid_t original_uid = geteuid();
+        if (original_uid == 0) {
+            // we need to drop root privileges
+            // otherwise we have access to everything
+            // UID == 65534 is often assigned to the user "nobody"
+            AVS_UNIT_ASSERT_SUCCESS(seteuid(65534));
+        }
+
+        char name[] = "/tmp/locked-XXXXXX";
+        AVS_UNIT_ASSERT_NOT_NULL(mkdtemp(name));
+        int retval = chmod(name, 0);
+        if (retval) {
+            (void) rmdir(name);
+        }
+        AVS_UNIT_ASSERT_SUCCESS(retval);
+        const avs_net_trusted_cert_info_t no_permissions_dir =
+                avs_net_trusted_cert_info_from_path(name);
+        retval = _avs_net_mbedtls_load_ca_certs(&chain,
+                                                &no_permissions_dir);
+        (void) rmdir(name);
+        AVS_UNIT_ASSERT_FAILED(retval);
+
+        if (original_uid == 0) {
+            // restore root privileges if we had them
+            // we need _POSIX_SAVED_IDS feature for this to work
+            AVS_STATIC_ASSERT(_POSIX_SAVED_IDS, posix_saved_ids);
+            AVS_UNIT_ASSERT_SUCCESS(seteuid(0));
+        }
+    }
+
     free(chain);
 }
